# openapi.yaml
components:
  schemas:
    AgentResponse:
      properties:
        channel:
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        entity_id:
          title: Entity Id
          type: string
        entity_type:
          const: agent
          default: agent
          title: Entity Type
          type: string
        etag:
          title: Etag
          type: string
        meta:
          additionalProperties: true
          title: Meta
          type: object
        version_id:
          title: Version Id
          type: string
      required:
      - entity_id
      - version_id
      - channel
      - etag
      - meta
      - content
      title: AgentResponse
      type: object
...
    EntityCreateRequest:
      properties:
        channel:
          default: latest
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        embedding:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Embedding
        evolution_meta:
          anyOf:
          - $ref: '#/components/schemas/EvolutionMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/EntityMeta'
        parent_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Version Id
      required:
      - meta
      - content
      title: EntityCreateRequest
      type: object
    EntityMeta:
      properties:
        author:
          anyOf:
          - type: string
          - type: 'null'
          title: Author
        name:
          maxLength: 255
          minLength: 1
          title: Name
          type: string
        namespace:
          anyOf:
          - type: string
          - type: 'null'
          title: Namespace
        tags:
          items:
            type: string
          title: Tags
          type: array
        when_to_use:
          anyOf:
          - type: string
          - type: 'null'
          title: When To Use
      required:
      - name
      title: EntityMeta
      type: object
...
    EntityUpdateRequest:
      properties:
        change_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Change Summary
        channel:
          default: latest
          title: Channel
          type: string
        content:
          additionalProperties: true
          title: Content
          type: object
        embedding:
          anyOf:
          - items:
              type: number
            type: array
          - type: 'null'
          title: Embedding
        evolution_meta:
          anyOf:
          - $ref: '#/components/schemas/EvolutionMeta'
          - type: 'null'
        meta:
          anyOf:
          - $ref: '#/components/schemas/EntityMeta'
          - type: 'null'
        parent_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Version Id
      required:
      - content
      title: EntityUpdateRequest
      type: object
    EvolutionMeta:
      properties:
        behavioral_descriptors:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Behavioral Descriptors
        fitness:
          anyOf:
          - type: number
          - type: 'null'
          title: Fitness
        is_valid:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Valid
        metrics:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metrics
        mutation_kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Mutation Kind
        prompt_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Prompt Ref
      title: EvolutionMeta
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  description: 'Persistent memory for CARL artifacts: steps, chains, agents, memory
    cards'
  title: GigaEvo Memory Module
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/agents:
    get:
      description: List all agents with pagination.
      operationId: list_agents_v1_agents_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          title: Offset
          type: integer
      - in: query
        name: channel
        required: false
        schema:
          default: latest
          title: Channel
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AgentResponse'
                title: Response List Agents V1 Agents Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Agents
      tags:
      - agents
    post:
      description: Create a new agent entity with its first version.
      operationId: create_agent_v1_agents_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Agent
      tags:
      - agents
  /v1/agents/{agent_id}:
    delete:
      description: Soft-delete an agent.
      operationId: delete_agent_v1_agents__agent_id__delete
      parameters:
      - in: path
        name: agent_id
        required: true
        schema:
          format: uuid
          title: Agent Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Agent
      tags:
      - agents
    get:
      description: Get an agent by ID, resolving the channel to a specific version.
      operationId: get_agent_v1_agents__agent_id__get
      parameters:
      - in: path
        name: agent_id
        required: true
        schema:
          format: uuid
          title: Agent Id
          type: string
      - in: query
        name: channel
        required: false
        schema:
          default: latest
          title: Channel
          type: string
      - in: header
        name: if-none-match
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-None-Match
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Agent
      tags:
      - agents
    put:
      description: Update an agent by creating a new immutable version.
      operationId: update_agent_v1_agents__agent_id__put
      parameters:
      - in: path
        name: agent_id
        required: true
        schema:
          format: uuid
          title: Agent Id
          type: string
      - in: header
        name: if-match
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Agent
      tags:
      - agents
...